bank/requirements.txt
sanic
mysql-connector-python
simple-crypt